requirements.txt
# Data processing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
librosa>=0.10.0
soundfile>=0.12.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# RAG / LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0

# PDF extraction
PyMuPDF>=1.23.0

# Application
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
